More linting and update to 1.19 [#3146]
Closes #3146. Updates to 1.19 and adds linting.

- Updates to 1.19, which adds new formatting rules for documentation comments.
- Adds more lint rules, some suggested by Notional, some chosen by me.

Changelog: refactor

+++ internal/core/block/block_end.go
@@ -170,7 +170,7 @@ func (m *Executor) EndBlock(block *Block) error {
 	// Index the root chain
 	rootIndexIndex, err := addIndexChainEntry(ledger.RootChain().Index(), &protocol.IndexEntry{
 		Source:     uint64(rootChain.Height() - 1),
-		BlockIndex: uint64(block.Index),
+		BlockIndex: block.Index,
 		BlockTime:  &block.Time,
 	})
 	if err != nil {
@@ -708,7 +708,7 @@ func (x *Executor) buildDirectoryAnchor(block *Block, systemLedger *protocol.Sys
 	// cannot be populated until the next block starts.
 	anchor := new(protocol.DirectoryAnchor)
 	anchor.Source = x.Describe.NodeUrl()
-	anchor.MinorBlockIndex = uint64(block.Index)
+	anchor.MinorBlockIndex = block.Index
 	anchor.MajorBlockIndex = block.State.MakeMajorBlock
 	anchor.Updates = systemLedger.PendingUpdates
 	if block.State.Anchor.ShouldOpenMajorBlock {
@@ -770,7 +770,7 @@ func (x *Executor) buildPartitionAnchor(block *Block, ledger *protocol.SystemLed
 	// cannot be populated until the next block starts.
 	anchor := new(protocol.BlockValidatorAnchor)
 	anchor.Source = x.Describe.NodeUrl()
-	anchor.MinorBlockIndex = uint64(block.Index)
+	anchor.MinorBlockIndex = block.Index
 	anchor.MajorBlockIndex = block.State.MakeMajorBlock
 	anchor.AcmeBurnt = ledger.AcmeBurnt
 	return anchor, nil
